+++ src/main/java/com/thiru/investment_tracker/entity/Transaction.java
@@ -50,7 +50,7 @@ public class Transaction {
 	private double price;
 
 	@Field("quantity")
-	private Long quantity;
+	private Double quantity;
 
 	@Field("total_value")
 	private double totalValue;
